Readers sub-package in antiweb_lib
Created a sub-package for the readers
