+++ antiweb.py
@@ -59,28 +59,12 @@ Objects
    :py:class:`lines <Line>`. The :py:class:`document <Document>` process all
    :py:class:`directives <Directives>`  to generate the output document.
 
-*******
-Readers
-*******
-
-Readers are responsible for the language dependent
-source parsing.
-
-@include(Reader doc)
-@include(CReader doc)
-@include(CSharpReader doc)
-@include(PythonReader doc)
-@include(ClojureReader doc)
-@include(GenericReader doc)
-@include(rstReader doc)
-
-
 ********
 Document
 ********
 
 @include(Document doc)
-@include(Line doc)
+@include(Line doc, antiweb_lib\readers\Reader.py)
 
 
 ***********
@@ -92,8 +76,6 @@ File Layout
 @include(imports)
 
 @include(management)
-
-@include(readers)
 
 @include(document)
 
@@ -235,18 +217,20 @@ Directives
 #@code
 from optparse import OptionParser
 import pygments.lexers as pm
-from pygments.token import Token
-import bisect
-import re
 import logging
 import sys
 import os.path
 import operator
 import os
 from sys import platform as _platform
-from bs4 import BeautifulSoup,  Tag,  NavigableString
 
-from antiweb_lib.directives import *
+from antiweb_lib.readers.Reader import Reader, Line
+from antiweb_lib.readers.CReader import CReader
+from antiweb_lib.readers.CSharpReader import CSharpReader
+from antiweb_lib.readers.ClojureReader import ClojureReader
+from antiweb_lib.readers.GenericReader import GenericReader
+from antiweb_lib.readers.rstReader import rstReader
+from antiweb_lib.readers.PythonReader import PythonReader
 
 #@rstart(management)
 
@@ -255,787 +239,13 @@ from antiweb_lib.directives import *
 
 #@code
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"
 
 logger = logging.getLogger('antiweb')
 
 class WebError(Exception):
     def __init__(self, error_list):
         self.error_list = error_list
-
-#@rstart(readers)
-#.. _readers:
-#
-
-#<<readers>>
-#===========
-
-#@code
-
-#@cstart(Reader)
-
-re_line_start = re.compile("^", re.M) #to find the line start indices
-
-
-class Reader(object):
-    #@start(Reader doc)
-    #Reader
-    #=============
-    """
-    .. py:class:: Reader(lexer)
-
-       This is the base class for all readers. The public functions
-       exposed to :py:class:`Document` are :py:meth:`process`,
-       and :py:meth:`filter_output`.
-
-       The main tasks for a reader is:
-
-         * Recognize lines that can contain directives. (comment lines or doc strings).
-         * Modify the source for language specific optimizations.
-         * Filter the processed output.
-       
-       :param lexer: A pygments lexer for the specified language
-    """
-    #@
-    #@indent 3
-    #@include(Reader)
-    #@include(Reader.__init__ doc)
-    #@include(Reader.process doc)
-    #@include(Reader.filter_output doc)
-    #@include(Reader._handle_token doc)
-    #@include(Reader._cut_comment doc)
-    #@include(Reader._post_process doc)
-    #@include(Reader._accept_token doc)
-    #@(Reader doc)
-    #Public Methods
-    #@cstart(Reader.__init__)
-    def __init__(self, lexer, single_comment_markers, block_comment_markers):
-        """
-        .. py:method:: __init__(lexer, single_comment_markers, block_comment_markers)
-            
-           The constructor initialises the language specific pygments lexer and comment markers. 
-           
-           :param Lexer lexer: The language specific pygments lexer.
-           :param string[] single_comment_markers: The language specific single comment markers.
-           :param string[] block_comment_markers: The language specific block comment markers.
-        """
-        self.lexer = lexer
-        self.single_comment_markers = single_comment_markers
-        self.block_comment_markers = block_comment_markers
-
-    #@cstart(Reader.process)
-    def process(self, fname, text):
-        """
-        .. py:method:: process(fname, text)
-
-           Reads the source code and identifies the directives.
-           This method is call by :py:class:`Document`.
-
-           :param string fname: The file name of the source code
-           :param string text: The source code
-           :return: A list of :py:class:`Line` objects.
-        """
-        text = text.replace("\t", " "*8)
-        starts = [ mo.start() for mo in re_line_start.finditer(text) ]
-        lines = [ Line(fname, i, l) for i, l in enumerate(text.splitlines()) ]
-
-        self.lines = lines    # A list of lines
-        self.starts = starts  # the start indices of the lines
-
-        tokens = self.lexer.get_tokens_unprocessed(text)
-        for index, token, value in tokens:
-            self._handle_token(index, token, value)
-        self._post_process(fname, text)
-        return self.lines
-
-
-    #@cstart(Reader.filter_output)
-    def filter_output(self, lines):
-        """
-        .. py:method:: filter_output(lines)
-
-           This method is call by :py:class:`Document` and gives
-           the reader the chance to influence the final output.
-        
-        """
-        return lines
-        
-    #@
-    
-    #Protected Methods
-    #@cstart(Reader._accept_token)
-    def _accept_token(self, token):
-        """
-        .. py:method:: _accept_token(token)
-
-           Checks if the token type may contain a directive.
-
-           :param token: A pygments token
-           :return: ``True`` if the token may contain a directive.
-                    ``False`` otherwise.
-        """
-        return True
-
-
-    #@cstart(Reader._post_process)
-    def _post_process(self, fname, text):
-        """
-        .. py:method:: _post_process(fname, text)
-
-           Does some post processing after the directives where found.
-        """
-
-        #correct the line attribute of directives, in case there have
-        #been lines inserted or deleted by subclasses of Reader
-        for i, l in enumerate(self.lines):
-            for d in l.directives:
-                d.line = i
-
-        #give the directives the chance to match
-        for l in self.lines:
-            for d in l.directives:
-                d.match(self.lines)
-
-
-    #@cstart(Reader._handle_token)
-    def _handle_token(self, index, token, value):
-        """
-        .. py:method:: _handle_token(index, token, value)
-
-           Find antiweb directives in valid pygments tokens.
-
-           :param integer index: The index within the source code
-           :param token: A pygments token.
-           :param string value: The token value.
-		"""
-		
-        if not self._accept_token(token): return
-        cvalue = self._cut_comment(index, token, value)
-        offset = value.index(cvalue)
-        for k, v in list(directives.items()):
-            for mo in v.expression.finditer(cvalue):
-                li = bisect.bisect(self.starts, index+mo.start()+offset)-1
-                line = self.lines[li]
-                line.directives = list(line.directives) + [ v(line.index, mo) ]
-         
-
-    #@cstart(Reader._cut_comment)
-    def _cut_comment(self, index, token, text):
-        """
-        .. py:method:: _cut_comment(index, token, value)
-
-           Cuts of the comment identifiers.
-
-           :param integer index: The index within the source code
-           :param token: A pygments token.
-           :param string value: The token value.
-           :return: value without comment identifiers.
-        """
-        return text
-
-#@cstart(CReader)
-
-class CReader(Reader):
-    #@start(CReader doc)
-    
-    #CReader
-    #=============
-    """
-    .. py:class:: CReader
-
-       A reader for C/C++ code. This class inherits :py:class:`Reader`.
-    """
-    #@indent 3
-    #@include(CReader)
-    #@(CReader doc)
-
-    def __init__(self, lexer,  single_comment_markers,  block_comment_markers):
-        super(CReader, self).__init__(lexer,  single_comment_markers,  block_comment_markers)
-        #For C/C++ exists only one type of single and block comment markers.
-        #The markers are retrieved here in order to avoid iterating over the markers every time they are used.
-        self.single_comment_marker = single_comment_markers[0]
-        self.block_comment_marker_start = block_comment_markers[0]   
-        self.block_comment_marker_end = block_comment_markers[1]
-
-    def _accept_token(self, token):
-        return token in Token.Comment
-    
-    def _cut_comment(self, index, token, text):
-        if text.startswith(self.block_comment_marker_start):
-            text = text[2:-2]
-    
-        elif text.startswith(self.single_comment_marker):
-            text = text[2:]
-
-        return text
-		
-    def filter_output(self, lines):
-        """
-        .. py:method:: filter_output(lines)
-
-           See :py:meth:`Reader.filter_output`.
-        """
-        print(lines)
-        for l in lines:
-            if l.type == "d":
-                #remove comment chars in document lines
-                stext = l.text.lstrip()
-
-                if stext == self.block_comment_marker_start or stext == self.block_comment_marker_end:
-                    #remove /* and */ from documentation lines
-                    #see the l.text.lstrip()! if the lines ends with a white space
-                    #the quotes will be kept! This is feature, to force the quotes
-                    #in the output
-                    continue
-                
-                if stext.startswith(self.single_comment_marker):
-                    l.text = l.indented(stext[2:])
-                            
-            yield l
-#@edoc
-
-#@cstart(CSharpReader)
-
-
-class CSharpReader(CReader):
-    #@start(CSharpReader doc)
-    #CSharpReader
-    #=============
-    """
-    .. py:class:: CSharpReader
-
-       A reader for C# code. This class inherits :py:class:`CReader`.
-       The CSharpReader is needed because C# specific output filtering is applied.
-       Compared to C, C# uses XML comments starting with *///* which are reused
-       for the final documentation. Here you can see an overview of the CSharpReader class
-       and its methods.
-       
-    """
-    
-    #@indent 3
-    #@include(CSharpReader)
-    #@include(CSharpReader.strip_tags doc)
-    #@include(CSharpReader.get_attribute_text doc)
-    #@include(CSharpReader.get_stripped_xml_lines doc)
-    #@include(CSharpReader.create_new_lines doc)
-    #@include(CSharpReader.init_xml_block doc)
-    #@include(CSharpReader.filter_output doc)
-    #@(CSharpReader doc)
-    
-    default_xml_block_index = -1
-    
-    def __init__(self, lexer,  single_comment_markers,  block_comment_markers):
-        #the call to the base class CReader is needed for initialising the comment markers
-        super(CSharpReader, self).__init__(lexer,  single_comment_markers,  block_comment_markers)
-            
-    #@cstart(CSharpReader.strip_tags)
-    def strip_tags(self,  tags):
-        """
-        .. py:method:: strip_tags(tags)
-
-           Removes all C# XML tags. The tags are replaced by their attributes and contents.
-           This method is called recursively. This is needed if the content of a tag also contains
-           tags.
-           
-           Examples:
-           
-            * ``<param name="arg1">value</param>`` will be stripped to : *"arg1 value"*
-            * ``<para>Start <see cref="(String)"/> end</para>`` will be stripped to : *"Start (String) end"*
-
-           :param Tag tags: The parsed xml tags.
-           :return: the XML tags replaced by their attributes and contents.
-        """
-        
-        if isinstance(tags,  Tag) and not tags.contents:            
-            return self.get_attribute_text(tags)
-    
-        for tag in tags.find_all(True):
-            text = self.get_attribute_text(tag)
-            
-            #if the content is a Navigablestring the content is simply concatenated 
-            #otherwise the contents of the content are recursively checked for xml tags
-            for content in tag.contents:
-                if not isinstance(content, NavigableString):
-                    content = self.strip_tags(content)
-                text += content
-    
-            tag.replaceWith(text)
-    
-        return tags
-    #@edoc
-    
-    #@cstart(CSharpReader.get_attribute_text)
-    def get_attribute_text(self,  tag):
-        """
-        .. py:method:: get_attribute_text(tag)
-
-           Returns the values of all XML tag attributes seperated by whitespace.
-
-           Examples:
-            
-            * ``<param name="arg1">parameterValue</param>`` returns : *"arg1 "*
-            * ``<include file='file1' path='[@name="test"]/*' />`` returns: *"file1 [@name="test"]/* "*
-
-           :param Tag tag: A BeautifulSoup Tag.
-           :return: the values of all attributes concatenated
-        """
-        #collect all values of xml tag attributes
-        attributes = tag.attrs
-        attribute_text = ""
-        for attribute, value in attributes.items():
-            attribute_text = value + " "
-        return attribute_text
-    #@edoc
-    
-    #@cstart(CSharpReader.get_stripped_xml_lines)
-    def get_stripped_xml_lines(self,  xml_lines_block):
-        """
-        .. py:method:: get_stripped_xml_lines(xml_lines_block)
-        
-           Removes all XML comment tags from the lines in the xml_lines_block.
-           
-           :param Line[] xml_lines_block: A list containing all Line object which belong to an XML comment block.
-           :return: A list of all stripped XML lines.
-        """
-        #the xml_lines_block contains Line objects. As the xml parser expects a string 
-        #the Line obects have to be converted to a string.
-        xml_text = "\n".join(map(operator.attrgetter("text"), xml_lines_block))
-        
-        #the xml lines will be parsed and then all xml tags will be removed       
-        xml_tags = BeautifulSoup(xml_text, "html.parser")
-        self.strip_tags(xml_tags)
-        stripped_xml_lines = xml_tags.text.splitlines()
-        return stripped_xml_lines
-    #@edoc
-    
-    #@cstart(CSharpReader.create_new_lines)
-    def create_new_lines(self, stripped_xml_lines, index, initial_line):
-        """
-        .. py:method:: create_new_lines(stripped_xml_lines, index, initial_line)
-        
-           This method is called after all XML comment tags are stripped. For each new 
-           line in the stripped_xml_lines a new Line object is created.
-           
-           :param string[] stripped_xml_lines: The comment lines were the XML tags have been stripped.
-           :param int index: The starting index for the new line object.
-           :param Line initial_line: The first line of the xml comment block. Its indentation will be used for all new created lines.
-           :return: A generator containing the lines which have been created out of the stripped_xml_lines.
-        """
-        for line in stripped_xml_lines:
-            new_line = Line(initial_line.fname,  index,  initial_line.indented(line.lstrip()))
-            index += 1
-            yield new_line
-    #@edoc
-   
-    #@cstart(CSharpReader.init_xml_block)
-    def init_xml_block(self):
-        """
-        .. py:method:: init_xml_block()
-        
-           Inits the variables which are needed for collecting an XML comment block.
-        """
-        xml_lines_block = []
-        xml_start_index = self.default_xml_block_index
-        return xml_lines_block,  xml_start_index
-    #@edoc
-    
-    #@cstart(CSharpReader.filter_output)
-    def filter_output(self, lines):
-        """
-        .. py:method:: filter_output(lines)
-        
-           Applies C# specific filtering for the final output.
-           XML comment tags are replaced by their attributes and contents.
-
-           See :py:meth:`Reader.filter_output`
-           
-           We have to handle four cases:
-            1. The current line is a code line: The line is added to result.
-            2. The current line is a block comment: The line can be skipped.
-            3. The current line is an XML comment: The line is added to the xml_lines_block.
-            4. The current line is a single comment line: Add the line to result. If the current line is the first
-               line after an xml comment block, the comment block is processed and its lines are added to result.
-           
-           :param Line[] lines: All lines of a file. The directives have already been replaced.
-           :return: A generator containing all lines for the final output.
-        """
-        #first the CReader filters the output
-        #afterwards the CSharpReader does some C# specific filtering
-        lines = super(CSharpReader, self).filter_output(lines)
-         
-        #the xml_lines_block collects xml comment lines
-        #if the end of an xml comment block is identified, the collected xml lines are processed
-        xml_lines_block, xml_start_index  = self.init_xml_block()
-        
-        for l in lines:
-            if l.type == "d":
-                #remove comment chars in document lines
-                stext = l.text.lstrip()
-                   
-                if stext == self.block_comment_marker_start or stext == self.block_comment_marker_end:
-                    #remove /* and */ from documentation lines. see the l.text.lstrip()!
-                    #if the lines ends with a white space the quotes will be kept! 
-                    #This is feature, to force the quotes in the output
-                    continue
-
-                if stext.startswith("/") and not stext.startswith(self.block_comment_marker_start):
-                    l.text = l.indented(stext[1:])
-                    
-                    if(xml_start_index == self.default_xml_block_index):
-                        #indicates that a new xml_block has started
-                        xml_start_index = l.index
-                        
-                    xml_lines_block.append(l)
-                    continue 
-                elif not xml_start_index == self.default_xml_block_index:
-                    #an xml comment block has ended, now the block is processed
-                    #at first the xml tags are stripped, afterwards a new line object is created for each 
-                    #stripped line and added to the final result generator
-                    stripped_xml_lines = self.get_stripped_xml_lines(xml_lines_block)
-                    
-                    new_lines = self.create_new_lines(stripped_xml_lines, xml_start_index, xml_lines_block[0])
-                    for line in new_lines:
-                        yield line
-                    
-                    #reset the xml variables for the next block
-                    xml_lines_block, xml_start_index  = self.init_xml_block()
-
-            yield l
-    #@edoc
-
-#@cstart(ClojureReader)
-class ClojureReader(Reader):
-    #@start(ClojureReader doc)
-    #ClojureReader
-    #=============
-    """
-    .. py:class:: ClojureReader
-
-       A reader for Clojure code. This class inherits :py:class:`Reader`.
-    """
-    #@indent 3
-    #@include(ClojureReader)
-    #@(ClojureReader doc)
-    def _accept_token(self, token):
-        return token in Token.Comment
-    
-    def _cut_comment(self, index, token, text):
-        if text.startswith(";"):
-            text = text[1:]
-        return text
-		
-    def filter_output(self, lines):
-        """
-        .. py:method:: filter_output(lines)
-
-           See :py:meth:`Reader.filter_output`.
-        """
-        for l in lines:
-            if l.type == "d":
-                #remove comment chars in document lines
-                stext = l.text.lstrip()
-		
-        if stext.startswith(";"):
-                    #remove comments but not chapters
-                    l.text = l.indented(stext[1:])
-
-        yield l
-    #@edoc
-    #@(ClojureReader)
-
-#@cstart(GenericReader)
-class GenericReader(Reader):
-    #@start(GenericReader doc)
-    #GenericReader
-    #=============
-    """
-    .. py:class:: GenericReader
-
-       A generic reader for languages with single line and block comments . 
-       This class inherits :py:class:`Reader`.
-    """
-    #@indent 3
-    #@include(GenericReader)
-    #@(GenericReader doc)
-    def __init__(self, single_comment_markers, block_comment_markers):
-        self.single_comment_markers = single_comment_markers
-        self.block_comment_markers = block_comment_markers
-
-    def _accept_token(self, token):
-        return token in Token.Comment
-    
-    def _cut_comment(self, index, token, text):
-        if text.startswith("/*"):
-            text = text[2:-2]
-    
-        elif text.startswith("//"):
-            text = text[2:]
-
-        return text
-
-    def filter_output(self, lines):
-
-        for l in lines:
-            if l.type == "d":
-                #remove comment chars in document lines
-                stext = l.text.lstrip()
-                for block_start, block_end in self.block_comment_markers: #comment layout: [("/*", "*/"),("#","@")]
-                    if stext == block_start or stext == block_end:
-                        #remove """ and ''' from documentation lines
-                        #see the l.text.lstrip()! if the lines ends with a white space
-                        #the quotes will be kept! This is feature, to force the quotes
-                        #in the output
-                        continue
-                    for comment_start in self.single_comment_markers: #comment layout: ["//",";"]
-                        if stext.startswith(comment_start):
-                            #remove comments but not chapters
-                            l.text = l.indented(stext[2:])
-                            
-            yield l
-
-
-#@cstart(rstReader)
-class rstReader(Reader):
-    #@start(rstReader doc)
-    #rstReader
-    #=============
-    """
-    .. py:class:: rstReader
-
-       A reader for rst code. This class inherits :py:class:`Reader`.
-    """
-    #@indent 3
-    #@include(rstReader)
-    #@(rstReader doc)
-    def _accept_token(self, token):
-        return token in Token.Comment
-    
-    def _cut_comment(self, index, token, text):
-        if text.startswith(".. "):
-            text = text[3:]
-
-        return text
-
-    def filter_output(self, lines):
-        """
-        .. py:method:: filter_output(lines)
-
-           See :py:meth:`Reader.filter_output`.
-        """
-        for l in lines:
-            if l.type == "d":
-                #remove comment chars in document lines
-                stext = l.text.lstrip()
-                if stext == '.. ':
-                    #remove """ and ''' from documentation lines
-                    #see the l.text.lstrip()! if the lines ends with a white space
-                    #the quotes will be kept! This is feature, to force the quotes
-                    #in the output
-                    continue
-                
-                if stext.startswith(".. "):
-                    #remove comments but not chapters
-                    l.text = l.indented(stext[3:])
-                            
-            yield l
-
-#@cstart(PythonReader)
-class PythonReader(Reader):
-    #@start(PythonReader doc)
-    #PythonReader
-    #============
-    '''
-    .. py:class:: PythonReader
-
-       A reader for python code. This class inherits :py:class:`Reader`.
-       To reduce the number of sentinels, the python reader does some more 
-       sophisticated source parsing:
-       
-       A construction like::
-         
-             @subst(_at_)cstart(foo)
-             def foo(arg1, arg2):
-                """ 
-                Foo's documentation
-                """ 
-                code
-
-
-       is replaced by::
-
-             @subst(_at_)cstart(foo)
-             def foo(arg1, arg2):
-                @subst(_at_)start(foo doc)
-                """ 
-                Foo's documentation
-                """ 
-                @subst(_at_)include(foo)
-                @subst(_at_)(foo doc)
-                code
-
-
-       The replacement will be done only:
-
-         * If the doc string begins with """
-         * If the block was started by a ``@rstart`` or ``@cstart`` directive
-         * If there is no antiweb directive in the doc string.
-         * Only a ``@cstart`` will insert the @include directive.
-
-
-       Additionally the python reader removes all single line ``"""`` and ``@subst(triple)``
-       from documentation lines. In the following lines::
-         
-             @subst(_at_)start(foo)
-             """ 
-             Documentation
-             """ 
-
-       The ``"""`` are automatically removed in the rst output. (see :py:meth:`filter_output`
-       for details).
-
-    '''
-
-    #@indent 3
-    #@include(PythonReader)
-    #@include(PythonReader._post_process doc)
-    #@include(PythonReader._accept_token doc)
-    #@include(PythonReader.filter_output doc)
-    #@include(PythonReader._cut_comment doc)
-    #@(PythonReader doc)
-    def __init__(self, lexer,  single_comment_markers,  block_comment_markers):
-        super(PythonReader, self).__init__(lexer,  single_comment_markers,  block_comment_markers)
-        self.doc_lines = []
-            
-    #@cstart(PythonReader._post_process)
-    def _post_process(self, fname, text):
-        """
-        .. py:method:: _post_process(fname, text)
-
-           See :py:meth:`Reader._post_process`.
-
-           This implementation *decorates* doc strings
-           with antiweb directives.
-        """
-        #from behind because we will probably insert some lines
-        self.doc_lines.sort(reverse=True)
-
-        #handle each found doc string
-        for start_line, end_line in self.doc_lines:
-            indents = set()
-
-            #@cstart(no antiweb directives in doc string)
-            #If antiweb directives are within the doc string,
-            #the doc string will not be decorated!
-            directives_between_start_and_end_line = False
-            for l in self.lines[start_line+1:end_line]:
-                if l:
-                    #needed for <<insert additional include>>
-                    indents.add(l.indent)
-                    
-                if l.directives:
-                    directives_between_start_and_end_line = True
-                    break
-
-            if directives_between_start_and_end_line: continue
-            
-            #@cstart(find the last directive before the doc string)
-            last_directive = None
-            for l in reversed(self.lines[:start_line]):
-                if l.directives:
-                    last_directive = l.directives[0]
-                    break
-            #@
-
-            if isinstance(last_directive, RStart):
-                #@cstart(decorate beginning and end)
-                l = self.lines[start_line]
-                start = Start(start_line, last_directive.name + " doc")
-                l.directives = list(l.directives) + [start]
-
-                l = self.lines[end_line]
-                end = End(end_line, last_directive.name + " doc")
-                l.directives = list(l.directives) + [end]
-                #@
-                
-                if isinstance(last_directive, CStart):
-                    #@cstart(insert additional include)
-                    l = l.like("")
-                    include = Include(end_line, last_directive.name)
-                    l.directives = list(l.directives) + [include]
-                    self.lines.insert(end_line, l)
-
-                    #the include directive should have the same 
-                    #indentation as the .. py:function:: directive 
-                    #inside the doc string. (It should be second 
-                    #value of sorted indents)
-                    indents = list(sorted(indents))
-                    if len(indents) > 1:
-                        l.change_indent(indents[1]-l.indent)
-                    #@
-
-        super(PythonReader, self)._post_process(fname, text)
-
-    #@edoc
-    #@rinclude(no antiweb directives in doc string)
-    #@rinclude(find the last directive before the doc string)
-    #@rinclude(decorate beginning and end)
-    #@rinclude(insert additional include)
-        
-    #@cstart(PythonReader._accept_token)
-    def _accept_token(self, token):
-        """
-        .. py:method:: _accept_token(token)
-
-           See :py:meth:`Reader._accept_token`.
-        """
-        return token in Token.Comment or token in Token.Literal.String.Doc
-
-
-    #@cstart(PythonReader._cut_comment)
-    def _cut_comment(self, index, token, text):
-        """
-        .. py:method:: _cut_comment(index, token, text)
-
-           See :py:meth:`Reader._cut_comment`.
-        """
-        if token in Token.Literal.String.Doc:
-            if text.startswith('"""'):
-                #save the start/end line of doc strings beginning with """
-                #for further decoration processing in _post_process,
-                start_line = bisect.bisect(self.starts, index)-1
-                end_line = bisect.bisect(self.starts, index+len(text)-3)-1
-                lines = list(filter(bool, text[3:-3].splitlines())) #filter out empty strings
-                if lines:
-                    self.doc_lines.append((start_line, end_line))
-                
-            text = text[3:-3]
-
-        return text
-
-    #@cstart(PythonReader.filter_output)
-    def filter_output(self, lines):
-        """
-        .. py:method:: filter_output(lines)
-
-           See :py:meth:`Reader.filter_output`.
-        """
-        for l in lines:
-            if l.type == "d":
-                #remove comment chars in document lines
-                stext = l.text.lstrip()
-
-                if stext == '"""' or stext == "'''":
-                    #remove """ and ''' from documentation lines
-                    #see the l.text.lstrip()! if the lines ends with a white space
-                    #the quotes will be kept! This is feature, to force the quotes
-                    #in the output
-                    continue
-                
-                if stext.startswith("#"):
-                    #remove comments but not chapters
-                    l.text = l.indented(stext[1:])
-                            
-            yield l
 
 # The keys are the lexer names of pygments
 readers = {
@@ -1044,7 +254,7 @@ readers = {
     "C#" : CSharpReader,
     "Python" : PythonReader,
     "Clojure" : ClojureReader,
-    "rst" : CReader,
+    "rst" : rstReader,
 }
 
 #@(readers)
@@ -1071,233 +281,6 @@ comments = {
 }
 #@
 
-#@cstart(Line)
-class Line(object):
-    #@start(Line doc)
-    #Line
-    #====
-    """
-    .. py:class:: Line(fname, index, text[, directives[, type]])
-
-       This class represents a text line.
-    """
-    #@indent 3
-    #@include(Line)
-    #@include(Line._directives doc)
-    #@include(Line.fname doc)
-    #@include(Line.index doc)
-    #@include(Line.text doc)
-    #@include(Line.type doc)
-    #@include(Line.indent doc)
-    #@include(Line.sindent doc)
-    #@include(Line.directives doc)
-    #@include(Line.directive doc)
-    #@include(Line.__init__ doc)
-    #@include(Line.set doc)
-    #@include(Line.like doc)
-    #@include(Line.indented doc)
-    #@include(Line.change_indent doc)
-    #@include(Line.__len__ doc)
-    #@include(Line.__repr__ doc)
-    #@(Line doc)
-    #Attributes
-    #@cstart(Line._directives)
-    _directives = ()
-    """
-    .. py:attribute:: _directives
-    
-       A list of :py:class:`Directive` objects, sorted
-       by their priority.
-    """
-    #@cstart(Line.fname)
-    fname = ""
-    """
-    .. py:attribute:: fname
-    
-       A string of the source's file name the line belongs to.
-    """
-    #@cstart(Line.index)
-    index = 0
-    """
-    .. py:attribute:: index
-    
-       The integer line index of the directive within the current block.
-    """
-    #@cstart(Line.text)
-    text = ""
-    """
-    .. py:attribute:: text
-    
-       A string containing the source line.
-    """
-    #@cstart(Line.type)
-    type = "d"
-    """
-    .. py:attribute:: type
-    
-       A char representing the line type:
-
-         * ``d`` stands for a document line
-         * ``c`` stands for a code line
-    """
-    #@
-
-    #Methods
-    #@cstart(Line.__init__)
-    def __init__(self, fname, index, text, directives=(), type='d'):
-        """
-        .. py:method:: __init__(name, index, text[, directives[, type]])
-
-           The constructor.
-        """
-        self.fname = fname
-        self.index = index
-        self.text = text
-        self.directives = directives
-        self.type = type
-
-
-    #@cstart(Line.set)
-    def set(self, index=None, type=None, directives=None):
-        """
-        .. py:method:: set([index=None[, type=None[, directives=None]]])
-
-           Changes the attributes :py:attr:`index`, :py:attr:`type`
-           and :py:attr:`directives` at once.
-
-           :param integer index: the line index.
-           :param char type: Either ``'d'`` or ``'c'``.
-           :param list directives: A list of :py:class:`DCirective` objects.
-           :return: The :py:class:`Line` object ``self``.
-        """
-        if index is not None:
-            self.index = index
-
-        if type is not None:
-            self.type = type
-
-        if directives is not None:
-            self.directives = directives
-
-        return self
-
-
-    
-    def clone(self, dline=None):
-
-        if dline is not None:
-            for d in self.directives:
-                d.line = dline
-
-        return Line(self.fname, self.index, self.text,
-                    self.directives[:], self.type)
-
-
-    #@cstart(Line.like)
-    def like(self, text):
-        """
-        .. py:method:: like(text)
-
-           Clones the Line with a different text.
-        """
-        return Line(self.fname, self.index, self.indented(text))
-
-
-    #@cstart(Line.indented)
-    def indented(self, text):
-        """
-        .. py:method:: indented(text)
-
-           Returns the text, with the same indentation as ``self``.
-        """
-        return self.sindent + text
-    
-    #@cstart(Line.change_indent)
-    def change_indent(self, delta):
-        """
-        .. py:method:: change_indent(delta)
-
-           Changes the lines indentation.
-        """
-        if delta < 0:
-            delta = min(-delta, self.indent)
-            self.text = self.text[delta:]
-
-        elif delta > 0:
-            self.text = " "*delta + self.text
-
-        return self
-
-    #@cstart(Line.__len__)
-    def __len__(self):
-        """
-        .. py:method:: __len__()
-
-           returns the length of the stripped :py:attr:`text`.
-        """
-        return len(self.text.strip())
-        
-
-    #@cstart(Line.__repr__)
-    def __repr__(self):
-        """
-        .. py:method:: __repr__()
-
-           returns a textual representation of the line.
-        """
-        return "Line(%i, %s, %s)" % (self.index, self.text, str(self.directives))
-    #@
-
-    #Properties
-    #@cstart(Line.indent)
-    @property
-    def indent(self):
-        """
-        .. py:attribute:: indent
-
-        An integer representing the line's indentation.
-        """
-        return len(self.text)-len(self.text.lstrip())
-
-
-    #@cstart(Line.sindent)
-    @property
-    def sindent(self):
-        """
-        .. py:attribute:: sindent
-
-        A string representation of the line's indentation.
-        """
-        return " "*self.indent
-
-
-    #@cstart(Line.directives)
-    @property
-    def directives(self):
-        """
-        .. py:attribute:: directives
-
-        A sorted sequence of :py:class:`Directive` objects.
-        """
-        return self._directives
-
-
-    @directives.setter
-    def directives(self, value):
-        self._directives = value[:]
-        if self._directives:
-            self._directives.sort(key=operator.attrgetter("priority"))
-
-
-    #@cstart(Line.directive)
-    @property
-    def directive(self):
-        """
-        .. py:attribute:: directive
-
-           The first of the contained :py:class:`Directive` objects.
-        """
-        return self.directives and self.directives[0]
      
 #@rstart(document)
